+++ src/model/Block.hpp
@@ -14,6 +14,7 @@
 #include <string>
 
 #include "global.hpp"
+#include "Decimal.hpp"
 
 namespace npge {
 
@@ -115,7 +116,7 @@ public:
     If a fragment doesn't have alignment row attached,
     then it is taken as is.
     */
-    double identity() const;
+    Decimal identity() const;
 
     /** Return consensus letter for given position.
     For each column, the most frequent letter is written to consensus.
