+++ src/util/AnyAs.hpp
@@ -61,13 +61,13 @@ public:
 
 /** Compare two any values.
 Any's must be of one of fillowing types:
-bool, int, double, string, vector<string>.
+bool, int, Decimal, string, vector<string>.
 */
 bool any_equal(const AnyAs& a, const AnyAs& b);
 
 /** Return if type of the option is good.
 Any's must be of one of fillowing types:
-bool, int, double, string, vector<string>.
+bool, int, Decimal, string, vector<string>.
 */
 bool good_opt_type(const std::type_info& ti);
